+++ frontEnd/crypto_wave/lib/repositories/wallet_repository/models/wallet_create.dart
@@ -11,7 +11,7 @@ class WalletCreate extends Equatable {
     required this.userId,
   });
 
-  final int currencyCount;
+  final double currencyCount;
   final String currencyName;
   final int userId;
